refactor: remove utils.c
